space2stats_api/requirements.txt
fastapi
uvicorn
pandas
pyarrow
s3fs
boto3
python-dotenv
shapely
h3
duckdb
pytest
httpx